feat(api/user): Move profiles to admin section
